# docs/swagger.yaml
basePath: /tasks
definitions:
  domain.Task:
    properties:
      checked:
        type: boolean
      id:
        type: integer
      text:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: b.abdusalamov@nv.dunice.net
    name: Batyr Abdusalamov
    url: https://twitter.com/
  description: Description of the ToDO REST API
  title: ToDo API
  version: "1.0"
paths:
  /tasks:
    delete:
      description: Responds 1 in JSON format.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: int
        "404":
          description: error
          schema:
            type: string
      summary: Delete all checked true tasks
      tags:
      - tasks
    get:
      description: Responds with the list of all tasks as JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Task'
            type: array
        "400":
          description: error
          schema:
            type: string
      summary: Get tasks array
      tags:
      - tasks
    patch:
      description: Responds 1 in JSON format.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: int
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Change checked task by ID
      tags:
      - tasks
    post:
      description: Responds 1 in JSON format.
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: int
        "400":
          description: error
          schema:
            type: string
      summary: Create task
      tags:
      - tasks
    put:
      description: Responds 1 in JSON format.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: int
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Change text task
      tags:
      - tasks
  /tasks/{checked}:
    patch:
      description: Responds 1 in JSON format.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: int
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Change checked all task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Responds 1 in JSON format.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: int
        "400":
          description: error
          schema:
            type: string
        "404":
          description: error
          schema:
            type: string
        "500":
          description: error
          schema:
            type: string
      summary: Delete single task by id
      tags:
      - tasks
swagger: "2.0"